dataclasses allocated to modules
