+++ manubot/cite/tests/test_citekey.py
@@ -7,8 +7,16 @@ from manubot.cite.citekey import (
     shorten_citekey,
     infer_citekey_prefix,
     inspect_citekey,
+    split_prefixed_identifier, 
+    CiteKey
 )
+from manubot.cite.doi import DOI
 
+def test_split_prefixed_identifier():    
+    assert split_prefixed_identifier("  @dOi:blah\t\t\n ") == ('doi', 'blah')
+
+def test_citekey_handle_results_in_doi(self):    
+    assert CiteKey("doi:blah").handle() == DOI("blah")
 
 @pytest.mark.parametrize("citation_string", [
     ('@doi:10.5061/dryad.q447c/1'),
