+++ manubot/cite/tests/test_doi.py
@@ -5,12 +5,9 @@ from manubot.cite.doi import (
     DOI
 )
 
-from manubot.cite.types import CiteKey
-
 
 class Test_DOI():
-    def test_citekey_handle_results_in_doi(self):    
-       assert CiteKey("doi:blah").handle() == DOI("blah")
+    pass
 
 
 def test_expand_short_doi():
